+++ types/src/commonMain/kotlin/kotools/types/NotEmptySet.kt
@@ -23,6 +23,11 @@ private constructor(private val elements: Set<E>) : Set<E> by elements {
                 ?: Result.failure(EmptyCollectionError)
     }
 
+    /** Returns this set as a [NotEmptyList]. */
+    @SinceKotools(Types, "4.0")
+    public fun toNotEmptyList(): NotEmptyList<E> =
+        elements.toNotEmptyListOrThrow()
+
     override fun toString(): String = "$elements"
 }