+++ types/src/commonTest/kotlin/kotools/types/NotEmptyListTest.kt
@@ -15,8 +15,7 @@ class NotEmptyListTest {
     fun toString_should_behave_like_a_Collection() {
         val collection: Collection<Int> = List(3) { NonZeroInt.random() }
             .map(NonZeroInt::toInt)
-        collection.toNotEmptyList()
-            .getOrThrow()
+        collection.toNotEmptyListOrThrow()
             .toString() assertEquals "$collection"
     }
 
@@ -56,8 +55,7 @@ class NotEmptyListTest {
     fun serialization_should_behave_like_a_List() {
         val list: List<Int> = List(3) { NonZeroInt.random() }
             .map(NonZeroInt::toInt)
-        val notEmptyList: NotEmptyList<Int> = list.toNotEmptyList()
-            .getOrThrow()
+        val notEmptyList: NotEmptyList<Int> = list.toNotEmptyListOrThrow()
         val result: String = Json.encodeToString(notEmptyList)
         result assertEquals Json.encodeToString(list)
     }